refactor: move components for components directory
