+++ app/src/components/InfoCard/index.tsx
