+++ app/src/components/WrapPage/index.tsx
@@ -1,16 +1,17 @@
-import { Outlet } from 'react-router-dom';
+import { Link, Outlet, useNavigate } from 'react-router-dom';
 import logo from '../../assets/logo-dynamox.png';
 import './wrapPage.css';
 
-function WrapPage() {
+export default function WrapPage() {
+  const navigate = useNavigate();
   return (
     <>
       <nav className="nav">
-        <img className="img" alt="Dynamox Logo" src={logo} />
+        <Link to="/" onClick={() => navigate('/')}>
+          <img className="img" alt="Dynamox Logo" src={logo} />
+        </Link>
       </nav>
       <Outlet />
     </>
   );
 }
-
-export default WrapPage;
