--- app/src/pages/Home/Footer/index.tsx
@@ -1,7 +1,0 @@
-export default function Footer() {
-  return (
-    <footer>
-      <h1>Footer</h1>
-    </footer>
-  );
-}
