+++ app/src/pages/Home/index.tsx
@@ -1,18 +1,26 @@
-import { Box, Button } from '@mui/material';
+import { Button, Typography } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
-import Footer from './Footer';
+import '../../index.css';
 
 function Home() {
   const navigate = useNavigate();
+
   return (
-    <>
-      <Box>
-        <Button onClick={() => navigate('/data')}>
-          Acessar dados da máquina
-        </Button>
-      </Box>
-      <Footer />
-    </>
+    <div className="home-container">
+      <Typography variant="h3">
+        Dynamox Front-end Developer Challenge
+      </Typography>
+      <Typography variant="h6">
+        Clique no botão abaixo para acessar a aba Análise de Dados.
+      </Typography>
+      <Button
+        onClick={() => navigate('/data')}
+        variant="contained"
+        sx={{ margin: '40px' }}
+      >
+        Acessar dados da máquina
+      </Button>
+    </div>
   );
 }
